Schema fixes
