+++ src/main/scala/org/constellation/primitives/CheckpointBlock.scala
@@ -2,11 +2,14 @@ package org.constellation.primitives
 
 import java.security.KeyPair
 
+import cats.data.{Ior, NonEmptyList, ValidatedNel}
 import constellation.signedObservationEdge
 import org.constellation.DAO
 import org.constellation.primitives.Schema._
 import org.constellation.util.HashSignature
 
+import cats.data._
+import cats.implicits._
 
 case class CheckpointBlock(
                             transactions: Seq[Transaction],
@@ -176,3 +179,212 @@ object CheckpointBlock {
   }
 
 }
+
+
+sealed trait CheckpointBlockValidation {
+  def errorMessage: String
+}
+
+case class EmptySignatures() extends CheckpointBlockValidation {
+  def errorMessage: String = "CheckpointBlock has no signatures"
+}
+
+
+case class InvalidSignature(signature: String) extends CheckpointBlockValidation {
+  def errorMessage: String = s"CheckpointBlock includes signature=$signature which is invalid"
+}
+
+object InvalidSignature {
+  def apply(s: HashSignature) = new InvalidSignature(s.signature)
+}
+
+
+case class InvalidTransaction(txHash: String) extends CheckpointBlockValidation {
+  def errorMessage: String = s"CheckpointBlock includes transaction=$txHash which is invalid"
+}
+
+object InvalidTransaction {
+  def apply(t: Transaction) = new InvalidTransaction(t.hash)
+}
+
+
+case class DuplicatedTransaction(txHash: String) extends CheckpointBlockValidation {
+  def errorMessage: String = s"CheckpointBlock includes duplicated transaction=$txHash"
+}
+
+object DuplicatedTransaction {
+  def apply(t: Transaction) = new DuplicatedTransaction(t.hash)
+}
+
+
+case class NoAddressCacheFound(txHash: String) extends CheckpointBlockValidation {
+  def errorMessage: String = s"CheckpointBlock includes transaction=$txHash which has no address cache"
+}
+
+object NoAddressCacheFound {
+  def apply(t: Transaction) = new NoAddressCacheFound(t.hash)
+}
+
+
+case class InsufficientBalance(address: String) extends CheckpointBlockValidation {
+  def errorMessage: String = s"CheckpointBlock includes transaction from address=$address which has insufficient balance"
+}
+
+object InsufficientBalance {
+  def apply(t: Transaction) = new InsufficientBalance(t.src.address)
+}
+
+
+// TODO: pass also a transaction metadata
+case class InternalInconsistency(cbHash: String) extends CheckpointBlockValidation {
+  def errorMessage: String = s"CheckpointBlock=$cbHash includes transaction/s which has insufficient balance"
+}
+
+object InternalInconsistency {
+  def apply(cb: CheckpointBlock) = new InternalInconsistency(cb.baseHash)
+}
+
+sealed trait CheckpointBlockValidatorNel {
+
+  type ValidationResult[A] = ValidatedNel[CheckpointBlockValidation, A]
+
+  def validateTransactionIntegrity(t: Transaction): ValidationResult[Transaction] =
+    if (t.valid) t.validNel else InvalidTransaction(t).invalidNel
+
+  def validateSourceAddressCache(t: Transaction)(implicit dao: DAO): ValidationResult[Transaction] =
+    dao.addressService
+      .get(t.src.address)
+      .fold[ValidationResult[Transaction]](NoAddressCacheFound(t).invalidNel)(_ => t.validNel)
+
+  def validateTransaction(t: Transaction)(implicit dao: DAO): ValidationResult[Transaction] =
+    validateTransactionIntegrity(t)
+      .product(validateSourceAddressCache(t))
+      .map(_ => t)
+
+  def validateTransactions(t: Iterable[Transaction])(implicit dao: DAO): ValidationResult[List[Transaction]] =
+    t.toList.map(validateTransaction(_).map(List(_))).combineAll
+
+  def validateDuplicatedTransactions(t: Iterable[Transaction]): ValidationResult[List[Transaction]] = {
+    val diff = t.toList.diff(t.toSet.toList)
+
+    if (diff.isEmpty) {
+      t.toList.validNel
+    } else {
+      def toError(t: Transaction): ValidationResult[Transaction] = DuplicatedTransaction(t).invalidNel
+
+      diff.map(toError(_).map(List(_))).combineAll
+    }
+  }
+
+  def validateSignatureIntegrity(s: HashSignature, baseHash: String): ValidationResult[HashSignature] =
+    if (s.valid(baseHash)) s.validNel else InvalidSignature(s).invalidNel
+
+  def validateSignature(s: HashSignature, baseHash: String): ValidationResult[HashSignature] =
+    validateSignatureIntegrity(s, baseHash)
+      .map(_ => s)
+
+  def validateSignatures(s: Iterable[HashSignature], baseHash: String): ValidationResult[List[HashSignature]] =
+    s.toList.map(validateSignature(_, baseHash).map(List(_))).combineAll
+
+  def validateEmptySignatures(s: Iterable[HashSignature]): ValidationResult[List[HashSignature]] =
+    if (s.nonEmpty) s.toList.validNel else EmptySignatures().invalidNel
+
+  def validateSourceAddressBalances(
+                                     t: Iterable[Transaction]
+                                   )(implicit dao: DAO): ValidationResult[List[Transaction]] = {
+    def lookup(key: String) = dao.addressService
+      .get(key)
+      .map(_.balance)
+      .getOrElse(0L)
+
+    def validateBalance(address: String, t: Iterable[Transaction]): ValidationResult[List[Transaction]] = {
+      val diff = lookup(address) - t.map(_.amount).sum
+
+      if (diff >= 0L) t.toList.validNel else InsufficientBalance(address).invalidNel
+    }
+
+    t.toList
+      .groupBy(_.src.address)
+      .map(a => validateBalance(a._1, a._2))
+      .toList
+      .combineAll
+  }
+
+  type AddressBalance = Map[String, Long]
+
+  def getParents(c: CheckpointBlock)(implicit dao: DAO): List[CheckpointBlock] =
+    c.parentSOEBaseHashes
+      .toList
+      .map(dao.checkpointService.get)
+      .map(_.flatMap(_.checkpointBlock))
+      .sequence[Option, CheckpointBlock]
+      .getOrElse(List())
+
+  def isInSnapshot(c: CheckpointBlock)(implicit dao: DAO): Boolean =
+    dao.threadSafeTipService
+      .acceptedCBSinceSnapshot
+      .contains(c.baseHash)
+
+  def getSummaryBalance(c: CheckpointBlock)(implicit dao: DAO): AddressBalance = {
+    val spend = c.transactions
+      .groupBy(_.src.address)
+      .mapValues(_.map(-_.amount).sum)
+
+    val received = c.transactions
+      .groupBy(_.dst.address)
+      .mapValues(_.map(_.amount).sum)
+
+    spend |+| received
+  }
+
+  /*
+      def getSnapshotBalances(implicit dao: DAO): AddressBalance =
+        dao.threadSafeTipService
+          .getSnapshotInfo()
+          .addressCacheData
+          .mapValues(_.balanceByLatestSnapshot)
+  */
+
+  def validateDiff(a: (String, Long))(implicit dao: DAO): Boolean = a match {
+    case (hash, diff) => dao.addressService.get(hash).map{_.balanceByLatestSnapshot}.getOrElse(0L) + diff >= 0
+  }
+
+  def validateCheckpointBlockTree(cb: CheckpointBlock)(implicit dao: DAO): Ior[NonEmptyList[CheckpointBlockValidation], AddressBalance] =
+    if (isInSnapshot(cb)) Map.empty[String, Long].rightIor
+    else
+      getParents(cb)
+        .map(validateCheckpointBlockTree)
+        .foldLeft(Map.empty[String, Long].rightIor[NonEmptyList[CheckpointBlockValidation]])((result, d) =>
+          result.combine(d))
+        .map(getSummaryBalance(cb) |+| _)
+        .flatMap(diffs =>
+          if (diffs.forall(validateDiff))
+            diffs.rightIor
+          else
+            Ior.both(NonEmptyList.of(InternalInconsistency(cb)), diffs))
+
+  implicit def validateTreeToValidated(v: Ior[NonEmptyList[CheckpointBlockValidation], AddressBalance]): ValidationResult[AddressBalance] =
+    v match {
+      case Ior.Right(a) => a.validNel
+      case Ior.Left(a) => a.invalid
+      case Ior.Both(a, _) => a.invalid
+    }
+
+  def validateCheckpointBlock(
+                               cb: CheckpointBlock
+                             )(implicit dao: DAO): ValidationResult[CheckpointBlock] = {
+    val preTreeResult =
+      validateEmptySignatures(cb.signatures)
+        .product(validateSignatures(cb.signatures, cb.baseHash))
+        .product(validateTransactions(cb.transactions))
+        .product(validateDuplicatedTransactions(cb.transactions))
+        .product(validateSourceAddressBalances(cb.transactions))
+
+    val postTreeIgnoreEmptySnapshot = if (dao.threadSafeTipService.lastSnapshotHeight == 0) preTreeResult
+    else preTreeResult.product(validateCheckpointBlockTree(cb))
+
+    postTreeIgnoreEmptySnapshot.map(_ => cb)
+  }
+}
+
+object CheckpointBlockValidatorNel extends CheckpointBlockValidatorNel
